pyproject.toml
[tool.poetry]
name = "samcam"
version = "0.1.0"
description = ""
authors = ["dnl-strnl <129630837+dnl-strnl@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "samcam", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.84"
numpy = "^2.2.1"
flask = "^3.1.0"
hydra-core = "^1.3.2"
"SAM-2" = { path = "./lib/segment-anything-2-real-time", develop = true }

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
